+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - fixed an issue with `Databricks Connection Manager` wich was broken due to some changes in the Databricks extension ([#143](/../../issues/143))
 - fix issue with named parameters in widgets ([#128](/../../issues/128))
 - minor fix to handle `pip install` more generic
+- removed support for code cells which never really worked
 
 **v2.0.3:**
 - fixed bug with `Buffer` introduced in `v2.0.2`